feat: add average rating for movie

+++ backend/services/movie-shelf-backend-core/src/main/java/com/radiuk/movieshelfbackendcore/dto/RatingDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 @NoArgsConstructor
 public class RatingDto {
 
-    private Byte score;
+    private Short score;
 }
