+++ backend/services/movie-shelf-backend-core/src/main/java/com/radiuk/movieshelfbackendcore/model/Rating.java
@@ -32,7 +32,7 @@ public class Rating {
     @JoinColumn(name = "movie_id")
     private Movie movie;
 
-    private Byte score;
+    private Short score;
 
     @NotNull
     @Column(name = "created_at")
